height point now in float

+++ jcrpg/src/org/jcrpg/world/place/geography/Forest.java
@@ -48,9 +48,9 @@ public class Forest extends Geography {
 	static Side[][] FOREST = new Side[][] { null, null, null,null,null,{new Side(TYPE_FOREST,SUBTYPE_FOREST)} };
 
 	@Override
-	protected int getPointHeightInside(int x, int z, int sizeX, int sizeZ, int worldX, int worldZ, boolean farView)
+	protected float getPointHeightInside(int x, int z, int sizeX, int sizeZ, int worldX, int worldZ, boolean farView)
 	{
-		Integer overrideHeight = overrideHeightForException(worldX, 0, worldZ, farView);
+		Float overrideHeight = overrideHeightForException(worldX, 0, worldZ, farView);
 		if (overrideHeight!=null) return overrideHeight;
 		//if (x<0 || z<0 || x>=sizeX || z>=sizeZ) return 0;
 		int Y = 0;
